requirements.txt
absl-py==0.11.0
requests==2.25.1
prometheus_client==0.9.0
cachetools==4.2.0
